# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blossomer GTM API v2",
...
    "contact": {
      "name": "Blossomer GTM API Support",
      "email": "support@blossomer.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.blossomer.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/accounts": {
      "post": {
        "tags": [
          "Accounts",
          "Accounts",
          "AI"
        ],
        "summary": "Generate Target Account Profile (discovery call preparation)",
        "description": "Generate a target account profile for authenticated users (Stack Auth JWT required).",
        "operationId": "prod_generate_target_account_api_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A structured discovery call preparation report with company analysis and ICP hypothesis.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetAccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts-crud": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Create Account",
        "description": "Create a new account for a company.\n\n- **company_id**: Company ID (must be owned by authenticated user)\n- **name**: Account name (required, max 255 chars)\n- **account_data**: JSON data with firmographics, buying signals, rationale",
        "operationId": "create_account_api_accounts_crud_post",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Company ID to create account for",
              "title": "Company Id"
            },
            "description": "Company ID to create account for"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get Accounts",
        "description": "Get all accounts for a company.\n\n- **company_id**: Company ID (must be owned by authenticated user)\n- **skip**: Number of accounts to skip (for pagination)\n- **limit**: Maximum number of accounts to return (1-1000)",
        "operationId": "get_accounts_api_accounts_crud_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Company ID to get accounts for",
              "title": "Company Id"
            },
            "description": "Company ID to get accounts for"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of accounts to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of accounts to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of accounts to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of accounts to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountResponse"
                  },
                  "title": "Response Get Accounts Api Accounts Crud Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts-crud/{account_id}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get Account",
        "description": "Get a specific account by ID.\n\nOnly returns accounts owned by the authenticated user (via company).",
        "operationId": "get_account_api_accounts_crud__account_id__get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Accounts"
        ],
        "summary": "Update Account",
        "description": "Update an account.\n\nOnly updates accounts owned by the authenticated user (via company).\nAll fields are optional - only provided fields will be updated.",
        "operationId": "update_account_api_accounts_crud__account_id__put",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Accounts"
        ],
        "summary": "Delete Account",
        "description": "Delete an account and all related data.\n\nOnly deletes accounts owned by the authenticated user (via company).\nThis will cascade delete all personas and campaigns.",
        "operationId": "delete_account_api_accounts_crud__account_id__delete",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts-crud/{account_id}/relations": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get Account With Relations",
        "description": "Get an account with all related personas and campaigns.\n\nOnly returns accounts owned by the authenticated user (via company).\nIncludes nested persona and campaign data.",
        "operationId": "get_account_with_relations_api_accounts_crud__account_id__relations_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountWithRelations"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "account_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Account Data",
            "description": "Account data including firmographics, buying signals, rationale, metadata"
          }
        },
        "type": "object",
        "required": [
          "name",
          "account_data"
        ],
        "title": "AccountCreate",
        "description": "Schema for creating a new account.",
        "example": {
          "account_data": {
            "buying_signals": [
              {
                "description": "Companies actively hiring developers indicating growth",
                "detection_method": "LinkedIn job postings, company announcements",
                "priority": "High",
                "title": "Recent engineering hiring",
                "type": "Company Data"
              },
              {
                "description": "Recent adoption of modern development tools",
                "detection_method": "GitHub repos, job descriptions, tech stack data",
                "priority": "Medium",
                "title": "DevOps tool adoption",
                "type": "Tech Stack"
              }
            ],
            "firmographics": {
              "employees": "50-500",
              "funding_stage": [
                "Series A",
                "Series B",
                "Series C"
              ],
              "geography": [
                "North America",
                "Europe"
              ],
              "industry": [
                "Software",
                "SaaS",
                "Technology"
              ],
              "keywords": [
                "rapid growth",
                "scaling team",
                "CI/CD",
                "automation",
                "developer productivity"
              ],
              "revenue": "$5M-$50M"
            },
            "rationale": [
              "Mid-market companies have complex workflows but limited resources",
              "Growing teams need better automation to maintain velocity",
              "Budget available for tools that improve developer productivity"
            ]
          },
          "name": "Mid-market SaaS Companies"
        }
      },
      "AccountResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "account_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Account Data",
            "description": "Account data including firmographics, buying signals, rationale, metadata"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Company Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "account_data",
          "id",
          "company_id",
          "created_at",
          "updated_at"
        ],
        "title": "AccountResponse",
        "description": "Schema for account responses."
      },
      "AccountUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "account_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Data"
          }
        },
        "type": "object",
        "title": "AccountUpdate",
        "description": "Schema for updating account information."
      },
      "AccountWithRelations": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "account_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Account Data",
            "description": "Account data including firmographics, buying signals, rationale, metadata"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Company Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "personas": {
            "items": {
              "$ref": "#/components/schemas/PersonaResponse"
            },
            "type": "array",
            "title": "Personas",
            "default": []
          },
          "campaigns": {
            "items": {
              "$ref": "#/components/schemas/CampaignResponse"
            },
            "type": "array",
            "title": "Campaigns",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "account_data",
          "id",
          "company_id",
          "created_at",
          "updated_at"
        ],
        "title": "AccountWithRelations",
        "description": "Account schema with related personas and campaigns."
      },
...
      "BuyingSignal": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Concise signal name (3-5 words)"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "1-2 sentences explaining why this signal indicates buying readiness"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Company Data|Website|Tech Stack|News|Social Media|Other"
          },
          "priority": {
            "$ref": "#/components/schemas/PriorityEnum",
            "description": "Low|Medium|High"
          },
          "detection_method": {
            "type": "string",
            "title": "Detection Method",
            "description": "Specific Clay enrichment or data source"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "type",
          "priority",
          "detection_method"
        ],
        "title": "BuyingSignal"
      },
...
      "CampaignResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "campaign_type": {
            "type": "string",
            "maxLength": 50,
            "title": "Campaign Type",
            "description": "Campaign type: email, linkedin, cold_call, ad"
          },
          "campaign_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Campaign Data",
            "description": "Campaign data including subject_line, content, segments, alternatives, configuration"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Account Id"
          },
          "persona_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Persona Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "campaign_type",
          "campaign_data",
          "id",
          "account_id",
          "persona_id",
          "created_at",
          "updated_at"
        ],
        "title": "CampaignResponse",
        "description": "Schema for campaign responses."
      },
...
      "Firmographics": {
        "properties": {
          "industry": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Industry",
            "description": "Exact industry names from Clay taxonomy"
          },
          "employees": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employees",
            "description": "Exact range (e.g., '50-500')"
          },
          "department_size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Size",
            "description": "Relevant dept size if applicable"
          },
          "revenue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue",
            "description": "Revenue range if relevant"
          },
          "geography": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geography",
            "description": "Geographic markets if relevant"
          },
          "business_model": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Model",
            "description": "Clay-searchable business model keywords"
          },
          "funding_stage": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Funding Stage",
            "description": "Exact funding stage names"
          },
          "company_type": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Type",
            "description": "Public/Private/PE-backed etc."
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords",
            "description": "3-5 sophisticated company description keywords that indicate implicit need - avoid obvious solution terms"
          }
        },
        "type": "object",
        "required": [
          "industry",
          "keywords"
        ],
        "title": "Firmographics"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ICPMetadata": {
        "properties": {
          "primary_context_source": {
            "type": "string",
            "title": "Primary Context Source",
            "description": "user_input|company_context|website_content"
          },
          "sources_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources Used",
            "description": "List of context sources utilized"
          },
          "confidence_assessment": {
            "$ref": "#/components/schemas/ConfidenceAssessment",
            "description": "Confidence metrics"
          },
          "processing_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Notes",
            "description": "Any important notes about analysis approach"
          }
        },
        "type": "object",
        "required": [
          "primary_context_source",
          "sources_used",
          "confidence_assessment"
        ],
        "title": "ICPMetadata"
      },
...
      "PersonaResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "persona_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Persona Data",
            "description": "Persona data including demographics, use cases, buying signals, objections, goals"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Account Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "persona_data",
          "id",
          "account_id",
          "created_at",
          "updated_at"
        ],
        "title": "PersonaResponse",
        "description": "Schema for persona responses."
      },
...
      "TargetAccountRequest": {
        "properties": {
          "website_url": {
            "type": "string",
            "title": "Website Url",
            "description": "Company website or landing page URL"
          },
          "account_profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Profile Name",
            "description": "Name of the target account profile (e.g., 'Mid-market SaaS companies', 'Enterprise healthcare organizations')"
          },
          "hypothesis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hypothesis",
            "description": "User's hypothesis about why this account profile is ideal for the solution"
          },
          "additional_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Context",
            "description": "Additional user-provided context for target account generation"
          },
          "company_context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Context",
            "description": "Company context from previous endpoints (e.g., company/generate output)"
          }
        },
        "type": "object",
        "required": [
          "website_url"
        ],
        "title": "TargetAccountRequest"
      },
      "TargetAccountResponse": {
        "properties": {
          "target_account_name": {
            "type": "string",
            "title": "Target Account Name",
            "description": "Short descriptive name for this customer segment"
          },
          "target_account_description": {
            "type": "string",
            "title": "Target Account Description",
            "description": "1-2 sentences: who they are and why they need this solution"
          },
          "target_account_rationale": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Account Rationale",
            "description": "3-5 bullets explaining the overall logic behind these targeting filters"
          },
          "firmographics": {
            "$ref": "#/components/schemas/Firmographics",
            "description": "Clay-ready prospect filters"
          },
          "buying_signals": {
            "items": {
              "$ref": "#/components/schemas/BuyingSignal"
            },
            "type": "array",
            "title": "Buying Signals",
            "description": "Detectable buying signals with specific data sources"
          },
          "buying_signals_rationale": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Buying Signals Rationale",
            "description": "3-5 bullets explaining the overall logic behind these buying signal choices"
          },
          "metadata": {
            "$ref": "#/components/schemas/ICPMetadata",
            "description": "Analysis metadata and quality scores"
          }
        },
        "type": "object",
        "required": [
          "target_account_name",
          "target_account_description",
          "target_account_rationale",
          "firmographics",
          "buying_signals",
          "buying_signals_rationale",
          "metadata"
        ],
        "title": "TargetAccountResponse",
        "description": "Response model for the /customers/target_accounts endpoint\n(ICP analysis with Clay-ready filters)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}